Move eventSource to hook
